# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EDCOCR API",
    "description": "REST API for forensic-grade document OCR processing.",
    "version": "4.1.0"
  },
  "paths": {
...
    "/api/v1/review/queue": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "List Review Queue",
        "description": "List review queue items with optional filtering.\n\nBy default returns pending items. Use the ``status`` parameter to\nfilter by a specific status (pending, approved, rejected, reprocess).",
        "operationId": "list_review_queue_api_v1_review_queue_get",
        "parameters": [
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by review reason",
              "title": "Reason"
            },
            "description": "Filter by review reason"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (default: pending)",
              "title": "Status"
            },
            "description": "Filter by status (default: pending)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewQueueResponse"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/review/stats": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Review Stats",
        "description": "Return aggregate review queue statistics.",
        "operationId": "review_stats_api_v1_review_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewStatsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/review/rules": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "List Review Rules",
        "description": "List all configured exception routing rules and their status.",
        "operationId": "list_review_rules_api_v1_review_rules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/review/{review_id}": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Review Item",
        "description": "Get details for a specific review item.",
        "operationId": "get_review_item_api_v1_review__review_id__get",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewItemResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/review/{review_id}/decision": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Review Decision",
        "description": "Submit a review decision (approve, reject, or request reprocessing).",
        "operationId": "review_decision_api_v1_review__review_id__decision_post",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewItemResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewDecisionRequest": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "approved",
              "rejected",
              "reprocess"
            ],
            "title": "Status"
          },
          "reviewer": {
            "type": "string",
            "title": "Reviewer",
            "default": ""
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ReviewDecisionRequest",
        "description": "Body for submitting a review decision."
      },
      "ReviewItemResponse": {
        "properties": {
          "review_id": {
            "type": "string",
            "title": "Review Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "quality_classification": {
            "type": "string",
            "title": "Quality Classification"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reviewer": {
            "type": "string",
            "title": "Reviewer"
          },
          "decision_notes": {
            "type": "string",
            "title": "Decision Notes"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "reviewed_at": {
            "type": "string",
            "title": "Reviewed At"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "review_id",
          "job_id",
          "reason",
          "confidence",
          "quality_classification",
          "status",
          "reviewer",
          "decision_notes",
          "created_at",
          "reviewed_at"
        ],
        "title": "ReviewItemResponse",
        "description": "Single review queue item."
      },
      "ReviewQueueResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewItemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ReviewQueueResponse",
        "description": "Paginated list of review queue items."
      },
      "ReviewStatsResponse": {
        "properties": {
          "pending": {
            "type": "integer",
            "title": "Pending"
          },
          "approved": {
            "type": "integer",
            "title": "Approved"
          },
          "rejected": {
            "type": "integer",
            "title": "Rejected"
          },
          "reprocess": {
            "type": "integer",
            "title": "Reprocess"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "avg_review_seconds": {
            "type": "number",
            "title": "Avg Review Seconds",
            "default": 0.0
          },
          "oldest_pending": {
            "type": "string",
            "title": "Oldest Pending",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "pending",
          "approved",
          "rejected",
          "reprocess",
          "total"
        ],
        "title": "ReviewStatsResponse",
        "description": "Review queue aggregate statistics."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "api__models__ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication. Set via OCR_API_KEY environment variable."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OAuth2 Bearer token (when OAUTH2_ENABLED=true)."
      }
    }
  },
...
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ]
}